requirements.txt
wheel==0.43.0
colored==2.2.4
colorama==0.4.6
langchain-openai==0.1.1
langchain==0.1.13
langchainhub==0.1.15
python-dotenv==1.0.1
python-box==7.1.1
playsound==1.2.2